# month_3/lesson_01(mentor)/api/docs/swagger.yaml
definitions:
  models.Branch:
    properties:
      address:
        type: string
      created_at:
        type: string
      founded_at:
        type: integer
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
      year:
        type: integer
    type: object
  models.CreateBranch:
    properties:
      address:
        type: string
      founded_at:
        type: integer
      name:
        type: string
    type: object
  models.CreateSales:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      payment_type:
        description: card, cash
        type: string
      price:
        type: number
      shop_assistant_id:
        type: string
      status:
        description: success, cancel
        type: string
    type: object
...
  models.GetAllBranch:
    properties:
      branches:
        items:
          $ref: '#/definitions/models.Branch'
        type: array
      count:
        type: integer
    type: object
  models.GetAllSalesResponse:
    properties:
      count:
        type: integer
      sales:
        items:
          $ref: '#/definitions/models.Sales'
        type: array
    type: object
...
  models.Sales:
    properties:
      branch_id:
        type: string
      cashier_id:
        type: string
      client_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      payment_type:
        description: card, cash
        type: string
      price:
        type: number
      shop_assistant_id:
        type: string
      status:
        description: success, cancel
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /branch:
    get:
      consumes:
      - application/json
      description: gets all branch based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllBranch'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL BRANCHS
      tags:
      - BRANCH
    post:
      consumes:
      - application/json
      description: CREATES BRANCH BASED ON GIVEN DATA
      parameters:
      - description: branch data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateBranch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES BRANCH
      tags:
      - BRANCH
...
  /sale:
    get:
      consumes:
      - application/json
      description: gets all sales based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllSalesResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL SALES
      tags:
      - SALE
    post:
      consumes:
      - application/json
      description: CREATES SALE BASED ON GIVEN DATA
      parameters:
      - description: branch data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateSales'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES SALE
      tags:
      - SALE
  /sale/{id}:
    delete:
      consumes:
      - application/json
      description: DELETES SALE BASED ON ID
      parameters:
      - description: id of sale
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE SALE BY ID
      tags:
      - SALE
    get:
      consumes:
      - application/json
      description: get sale by ID
      parameters:
      - description: Sale ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Sales'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - SALE
    put:
      consumes:
      - application/json
      description: UPDATES SALE BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of sale
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: sale data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.Sales'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE SALE BY ID
      tags:
      - SALE
...
swagger: "2.0"